Fixed Task 2 to not get any warning message upon compilation

+++ Lab 3/task2.cpp
@@ -20,7 +20,7 @@ class RationalNumber {
         }
 
         float convert() {
-            return numerator/denominator;
+            return (float)numerator/(float)denominator;
         }
 
         void invert(){
